Fixed bug in ASA and SAS
SAS: when searching for two pairs of congruent segments, then if only one pair was found, it would add the pair to the list of parents but would not remove it from the list when it went on two the next pair of congruent angles. Caused weird behavior.
ASA: same as above, swap angles with segments and vice versa